requirements.txt
django==4.1.4
cloudinary==1.30.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==0.21.0
drf-social-oauth2==1.2.1